Refactor and update song management and UI components
Refactored dependency injection and error handling for IDiscordRPC in App.xaml.cs. Updated route and singleton registrations to use SingleSongShellD instead of NowPlayingD. Changed access modifiers in ISongsManagementService. Improved playlist and song queue management using BehaviorSubject in PlaybackUtilsService. Updated project files to reflect new views and added folders. Removed NowPlayingD files and added SingleSongShellD with similar functionality. Enhanced HomePageVM and HomeD for better state management and UI responsiveness.

+++ Dimmer/DataAccess/Services/PlayListManagementService.cs
@@ -188,17 +188,22 @@ public class PlayListManagementService : IPlaylistManagementService
                     //Shell.Current.DisplayAlert("Error", "Playlist not found", "OK");
                     //return false;
                 }
-                var existingLink = db.All<PlaylistSongLink>()
-                    .FirstOrDefault(link => link.PlaylistId == specificPlaylist.Id && link.SongId == song.Id);
-
-                if (existingLink is not null)
+                var existingLinkList = db.All<PlaylistSongLink>()
+                .Where(link => link.PlaylistId == specificPlaylist.Id && link.SongId == song.Id)
+                .ToList();
+                    
+                if (existingLinkList is not null)
                 {
-                    db.Remove(existingLink);
-                }
+                    var existingLink = existingLinkList.First();
+                    if (existingLink is not null)
+                    {
+                        db.Remove(existingLink);
+                    }
 
-                specificPlaylist.TotalDuration -= song.DurationInSeconds;
-                specificPlaylist.TotalSize -= song.FileSize;
-                specificPlaylist.TotalSongsCount -= 1;
+                    specificPlaylist.TotalDuration -= song.DurationInSeconds;
+                    specificPlaylist.TotalSize -= song.FileSize;
+                    specificPlaylist.TotalSongsCount -= 1;
+                }
             });
             //GetPlayLists();
             return true;
